requirements.txt
matplotlib==1.3.1
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.0
tensorflow==1.12.0
